moved all of the mnist code into the init file like the other datasets
